added log4j latency benchmark
