+++ gflog-benchmark/src/main/java/com/epam/deltix/gflog/benchmark/util/Benchmark.java
@@ -1,0 +1,36 @@
+package com.epam.deltix.gflog.benchmark.util;
+
+import java.util.function.Consumer;
+
+
+public final class Benchmark {
+
+    private final String name;
+    private final Runnable prepare;
+    private final Runnable cleanup;
+    private final Consumer<BenchmarkState> command;
+
+    public Benchmark(final String name, final Runnable prepare, final Runnable cleanup, final Consumer<BenchmarkState> command) {
+        this.name = name;
+        this.prepare = prepare;
+        this.cleanup = cleanup;
+        this.command = command;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public Runnable getPrepare() {
+        return prepare;
+    }
+
+    public Runnable getCleanup() {
+        return cleanup;
+    }
+
+    public Consumer<BenchmarkState> getCommand() {
+        return command;
+    }
+
+}
