+++ gflog-benchmark/src/main/resources/com/epam/deltix/gflog/benchmark/gflog-noop-benchmark.xml
