+++ gflog-benchmark/src/main/resources/com/epam/deltix/gflog/benchmark/log4j/log4j-file-benchmark.xml
