+++ gflog-benchmark/src/main/resources/com/epam/deltix/gflog/benchmark/log4j/log4j-noop-benchmark.xml
